services/Gacha/requirements.txt
sqlmodel==0.0.22
fastapi==0.115.5
python-dotenv==1.0.1
uvicorn==0.32.1
pymysql==1.1.1
requests==2.32.3
pyjwt==2.10.1
python-multipart==0.0.19